backend/requirements.txt
python-slugify>=8.0.1
pytest>=7.4.3
